main-app/backend/taxifleet-management/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
asyncpg
python-dotenv
boto3
